requirements.txt
onnxruntime==1.21.0

# Whisper
git+https://github.com/openai/whisper.git 
ffmpeg-python

# Web Server
flask-ml==0.2.5

# Pyannote dependencie
pyannote.core==5.0.0
pyannote.audio==3.3.2
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1

#database
sqlite3

